Code refactoring in console version of the contact agenda
Scout's Rule:
"Leave the world a bit better than before"
Programmer Rule:
"Leave the code a bit better than before"
So I tried clean my code.
It is a different and a clean way of writing the same thing.
Previously, the code had many if and elif statements to the functions of our agenda(show,create,exit,delete and edit) and that is boring to write.
So I replaced that with a dictionary of functions.
The input is the key number to a function.
The only function which I needed to write some more things was the edit function
The refactored code is in Agenda de contactos yt- Copia(2)
If you find some error or some thing that I could refactor comment.